Fixed issue #4 (Sim does not handle non-sequential mote id numbering)
Added list to store ids of motes that exist, when loading a topo, check if the node on each side of a connection is already in the list, if not, add it. Instead of performing operations meant for all nodes on a range from 1 to numMote+1, perform it on each entry of the nodeids list.